티켓양도 작업중

+++ ticket24/src/main/java/kr/spring/ticket/controller/TicketAjaxController.java
@@ -27,7 +27,7 @@ public class TicketAjaxController {
 	 *=============================*/
 	@RequestMapping("ticket/deleteFile")
 	@ResponseBody
-	public Map<String,String> processFile(int ticket_num, HttpSession session, HttpServletRequest request){
+	public Map<String,String> processFile(int ticket_num, int file_num, HttpSession session, HttpServletRequest request){
 		
 		Map<String,String> mapJson = new HashMap<String,String>();
 		
@@ -37,12 +37,24 @@ public class TicketAjaxController {
 		}else {
 			TicketVO vo = ticketService.selectTicket(ticket_num);
 			
-			ticketService.deleteFile(ticket_num);
-			FileUtil.removeFile(request, vo.getTicket_filename1());
-			FileUtil.removeFile(request, vo.getTicket_filename2());
-			FileUtil.removeFile(request, vo.getTicket_filename3());
-			FileUtil.removeFile(request, vo.getTicket_filename4());
+			Map<String,Integer> map = new HashMap<String,Integer>();
+			map.put("ticket_num", ticket_num);
+			map.put("file_num", file_num);
 			
+			ticketService.deleteFile(map);
+			if(file_num==1) {
+				FileUtil.removeFile(request, vo.getTicket_filename1());
+			}
+			if(file_num==2) {
+				FileUtil.removeFile(request, vo.getTicket_filename2());
+			}
+			
+			if(file_num==3) {
+				FileUtil.removeFile(request, vo.getTicket_filename3());
+			}
+			if(file_num==4) {
+				FileUtil.removeFile(request, vo.getTicket_filename4());
+			}
 			mapJson.put("result", "success");
 		}
